pyproject.toml
[project]
name = "pr-reviewer"
version = "0.1.0"
description = "AI-powered PR analysis tool with Supabase integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "mcp[cli]>=1.9.0",
    "notion-client>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "supabase>=2.3.0",
    "openai>=1.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
